Alert user when they use an incompatible version of NetLogo
Due to changes in the tabs "API" of NetLogo desktop, LevelSpace will
only work with NetLogo 6.2.0 and newer. But we want to release updates
to the extensions library that will include users on NetLogo 6.1.1 and
6.1.0, for whom this newer LS will *not* work even though the API
versions say they will (all at 6.1). So we'll warn the user that they
need to upgrade NetLogo or downgrade LS, rather than just blowing up.